# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Music Streaming API
  version: 1.0.0
  description: API documentation for the Music Streaming application.
servers:
  - url: http://localhost:3000/api
    description: Local development server
...
paths:
...
  /songs/{id}:
    get:
      summary: Get a song by ID
      tags: [Songs]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Song retrieved successfully
        '404':
          description: Song not found
    put:
      summary: Update a song by ID
      tags: [Songs]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                artist_id:
                  type: integer
                album_id:
                  type: integer
                genre:
                  type: string
                duration:
                  type: integer
                file_url:
                  type: string
                cover_image_url:
                  type: string
                release_date:
                  type: string
                  format: date
      responses:
        '200':
          description: Song updated successfully
        '404':
          description: Song not found
    delete:
      summary: Delete a song by ID
      tags: [Songs]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Song deleted successfully
        '404':
          description: Song not found
...
  /playlists/{playlistId}/songs/{songId}:
    delete:
      summary: Remove a song from a playlist
      tags: [Playlists]
      parameters:
        - in: path
          name: playlistId
          required: true
          schema:
            type: integer
          description: ID of the playlist
        - in: path
          name: songId
          required: true
          schema:
            type: integer
          description: ID of the song to remove
      responses:
        '200':
          description: Song removed from playlist successfully
        '404':
          description: Playlist or song not found
...